Move the Movie data in a proper folder
