+++ utils/db/MovieCommentAdapter.ts
@@ -1,5 +1,5 @@
 import {WithId} from "mongodb";
-import {Comment} from "../movies/Movie";
+import {Comment} from "../../data/movies/Movie";
 import {MongoMovieDetail} from "./MoviesRepository";
 
 export const toDomainComment = (mongoMovieDetail: WithId<MongoMovieDetail>): Comment => {
